Separate exploded mount from war type
