+++ server/src/main/java/org/jboss/as/server/deployment/Phase.java
@@ -174,13 +174,14 @@ public enum Phase {
     }
 
     // STRUCTURE
-    public static final int STRUCTURE_WAR_DEPLOYMENT_INIT               = 0x0000;
+    public static final int STRUCTURE_EXPLODED_MOUNT                    = 0x0000;
     public static final int STRUCTURE_MOUNT                             = 0x0001;
     public static final int STRUCTURE_MANIFEST                          = 0x0100;
     // must be before osgi
     public static final int STRUCTURE_JDBC_DRIVER                       = 0x0150;
     public static final int STRUCTURE_OSGI_MANIFEST                     = 0x0200;
     public static final int STRUCTURE_RAR                               = 0x0300;
+    public static final int STRUCTURE_WAR_DEPLOYMENT_INIT               = 0x0400;
     public static final int STRUCTURE_WAR                               = 0x0500;
     public static final int STRUCTURE_EAR_DEPLOYMENT_INIT               = 0x0600;
     public static final int STRUCTURE_EAR_APP_XML_PARSE                 = 0x0700;
